+++ NEWS.md
@@ -3,9 +3,8 @@
 ## Release V0.2.0
 
 - A completely new internal design of the package.
-- Implemented elastic buckling analysis.
-- Implemented free vibrational analysis.
 - Added separate types for conc. and dist. loads.
+- Added an option to easily run planar analysis using `solve(model, analysistype, planar = true)`, i.e., automatically restrict $u_{z}$, $\theta_{x}$, and $\theta_{y}$ DOFs.
 - Added a plotting extension for quick model prototyping.
 - Better pretty-printing of a model's components.
 - Added automatic report generation that can be invoked by using `generatereport()` function.